pbuf: Introduce PBUF_FLAG_DATA_VALID
`PBUF_FLAG_DATA_VALID` indicates that the L4 checksum within
a pbuf is already validated and further checks are not
necessary by the stack. For implementation simplicity
`PBUF_FLAG_DATA_VALID` can only be set once for a pbuf chain
with the heading pbuf.

Typically, this flag is set by netif drivers in virtual
environments for traffic that is already validated, e.g.,
by receiving it on a physical network card.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>